# Conflicts:
#	src/moai-box2d/MOAIBox2DDebugDraw.cpp
#	src/moai-sim/MOAIDraw.cpp
#	src/moai-sim/MOAIGfxDeviceBase.cpp
#	src/moai-sim/MOAIGfxDeviceStateCache.cpp
#	src/moai-sim/MOAIGfxResourceMgr.cpp
#	src/moai-sim/MOAIImage.cpp
#	src/moai-sim/MOAISelectionMesh.cpp
#	src/moai-sim/MOAIVectorUtil.cpp
#	src/moai-util/MOAIStreamAdapter.cpp
#	src/zl-gfx/ZLGfxRetained.cpp
#	src/zl-util/ZLColor.cpp
#	src/zl-util/ZLCopyOnWrite.cpp
#	src/zl-util/ZLLeanStack.h

+++ src/moai-core/MOAILogMessages.h
@@ -44,9 +44,9 @@ public:
 
 	//----------------------------------------------------------------//
 	static bool		CheckFileExists					( cc8* filename, lua_State* L = 0 );
-	static bool		CheckIndex						( u32 idx, u32 size, lua_State* L = 0 );
-	static bool		CheckIndexPlusOne				( u32 idx, u32 size, lua_State* L = 0 );
-	static bool		CheckReserve					( u32 idx, u32 size, lua_State* L = 0 );
+	static bool		CheckIndex						( size_t idx, size_t size, lua_State* L = 0 );
+	static bool		CheckIndexPlusOne				( size_t idx, size_t size, lua_State* L = 0 );
+	static bool		CheckReserve					( size_t idx, size_t size, lua_State* L = 0 );
 	static void		RegisterDefaultLogMessages		();
 	static void		RegisterLogMessageIDs			( MOAILuaState& state );
 };
