+++ src/moai-image-jpg/MOAIImageFormatJpg.cpp
@@ -47,7 +47,7 @@ static void _jpgSkipInputData ( j_decompress_ptr cinfo, long num_bytes ) {
 		src->pub.next_input_byte = ( JOCTET* )(( size_t )src->pub.next_input_byte + num_bytes );
 	}
 	else {
-		src->stream->Seek ( num_bytes - src->pub.bytes_in_buffer, SEEK_CUR );
+		src->stream->Seek (( long )( num_bytes - src->pub.bytes_in_buffer ), SEEK_CUR );
 		src->pub.bytes_in_buffer = src->stream->ReadBytes ( src->buffer, JPG_BUFFER_SIZE );
 		src->pub.next_input_byte = ( JOCTET* )src->buffer;
 	}
