+++ src/moai-sim/MOAIAnimCurveVec.cpp
@@ -86,7 +86,7 @@ ZLVec3D MOAIAnimCurveVec::GetCurveDelta () const {
 
 	ZLVec3D delta;
 
-	u32 size = this->mKeys.Size ();
+	u32 size = ( u32 )this->mKeys.Size ();
 	if ( size > 1 ) {
 		delta = this->mSamples [ size - 1 ];
 		delta.Sub ( this->mSamples [ 0 ]);
