+++ src/moai-sim/MOAIPartition.cpp
@@ -462,7 +462,7 @@ u32 MOAIPartition::AffirmInterfaceMask ( u32 typeID ) {
 
 	u32 mask = this->GetInterfaceMask ( typeID );
 	if ( mask == 0 ) {
-		u32 top = this->mInterfaceIDs.Size ();
+		size_t top = this->mInterfaceIDs.Size ();
 		if ( top < INTERFACE_MASK_BITS ) {
 			this->mInterfaceIDs.Grow ( top + 1 );
 			this->mInterfaceIDs [ top ] = typeID;
@@ -484,8 +484,8 @@ void MOAIPartition::AffirmPriority ( MOAIProp& prop ) {
 //----------------------------------------------------------------//
 void MOAIPartition::Clear () {
 
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].Clear ();
 	}
 	this->mBiggies.Clear ();
@@ -498,8 +498,8 @@ u32 MOAIPartition::GatherProps ( MOAIPartitionResultBuffer& results, MOAIProp* i
 	
 	results.Reset ();
 	
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].GatherProps ( results, ignoreProp, interfaceMask, queryMask );
 	}
 	this->mBiggies.GatherProps ( results, ignoreProp, interfaceMask, queryMask );
@@ -514,8 +514,8 @@ u32 MOAIPartition::GatherProps ( MOAIPartitionResultBuffer& results, MOAIProp* i
 	
 	results.Reset ();
 	
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].GatherProps ( results, ignoreProp, point, orientation, interfaceMask, queryMask );
 	}
 	this->mBiggies.GatherProps ( results, ignoreProp, point, orientation, interfaceMask, queryMask );
@@ -529,8 +529,8 @@ u32 MOAIPartition::GatherProps ( MOAIPartitionResultBuffer& results, MOAIProp* i
 	
 	results.Reset ();
 	
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].GatherProps ( results, ignoreProp, point, this->mPlaneID, interfaceMask, queryMask );
 	}
 	this->mBiggies.GatherProps ( results, ignoreProp, point, interfaceMask, queryMask );
@@ -545,8 +545,8 @@ u32 MOAIPartition::GatherProps ( MOAIPartitionResultBuffer& results, MOAIProp* i
 	results.Reset ();
 	box.Bless ();
 	
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].GatherProps ( results, ignoreProp, box, this->mPlaneID, interfaceMask, queryMask );
 	}
 	this->mBiggies.GatherProps ( results, ignoreProp, box, interfaceMask, queryMask );
@@ -560,8 +560,8 @@ u32 MOAIPartition::GatherProps ( MOAIPartitionResultBuffer& results, MOAIProp* i
 	
 	results.Reset ();
 	
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].GatherProps ( results, ignoreProp, frustum, this->mPlaneID, interfaceMask, queryMask );
 	}
 	this->mBiggies.GatherProps ( results, ignoreProp, frustum, interfaceMask, queryMask );
@@ -573,8 +573,8 @@ u32 MOAIPartition::GatherProps ( MOAIPartitionResultBuffer& results, MOAIProp* i
 //----------------------------------------------------------------//
 u32 MOAIPartition::GetInterfaceMask ( u32 typeID ) const {
 
-	u32 total = this->mInterfaceIDs.Size ();
-	for ( u32 i = 0; i < total; ++i ) {
+	size_t total = this->mInterfaceIDs.Size ();
+	for ( size_t i = 0; i < total; ++i ) {
 		if ( this->mInterfaceIDs [ i ] == typeID ) {
 			return 1 << i;
 		}
@@ -656,8 +656,8 @@ void MOAIPartition::OnPropUpdated ( MOAIProp& prop ) {
 // This moves all props to the 'empties' cell
 void MOAIPartition::PrepareRebuild () {
 
-	u32 totalLevels = this->mLevels.Size ();
-	for ( u32 i = 0; i < totalLevels; ++i ) {
+	size_t totalLevels = this->mLevels.Size ();
+	for ( size_t i = 0; i < totalLevels; ++i ) {
 		this->mLevels [ i ].ExtractProps ( this->mEmpties, 0 );
 	}
 	this->mBiggies.ExtractProps ( this->mEmpties, 0 );
@@ -763,8 +763,8 @@ void MOAIPartition::UpdateProp ( MOAIProp& prop, u32 status ) {
 		
 		MOAIPartitionLevel* level = 0;
 		
-		u32 totalLevels = this->mLevels.Size ();
-		for ( u32 i = 0; i < totalLevels; ++i ) {
+		size_t totalLevels = this->mLevels.Size ();
+		for ( size_t i = 0; i < totalLevels; ++i ) {
 			
 			MOAIPartitionLevel* testLevel = &this->mLevels [ i ];
