+++ src/zl-util/ZLBase64Reader.h
@@ -29,7 +29,7 @@ public:
 	//----------------------------------------------------------------//
 	static size_t		EstimateDecodedLength	( size_t encodedLength );
 	u32					GetCaps					();
-	int					SetCursor				( long offset );
+	int					SetCursor				( size_t offset );
 						ZLBase64Reader			();
 						~ZLBase64Reader			();
 };
